# backend/apidoc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Um título de API",
    "description": "Descrição da API"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": {
    "Bearer": []
  },
...
  "paths": {
    "/auth/paciente": {
      "post": {
        "tags": ["AUTHORIZATION"],
        "summary": "Autoriza e gera token",
        "description": "Devolve um token válido",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "auth/paciente",
            "description": "Objeto que descreve um autorização para Paciente",
            "schema": {
              "type": "object",
              "required": ["senha", "email"],
              "properties": {
                "email": { "type": "string", "example": "ana@ana.com" },
                "senha": { "type": "string", "example": "12345" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação bem sucedida"
          },
          "400": {
            "description": "Operação inválida"
          }
        }
      }
    },
    "/auth/profissional": {
      "post": {
        "tags": ["AUTHORIZATION"],
        "summary": "Autoriza e gera token",
        "description": "Devolve um token válido",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "auth/profissional",
            "description": "Objeto que descreve um autorização para Profissional",
            "schema": {
              "type": "object",
              "required": ["senha", "email", "cro"],
              "properties": {
                "email": { "type": "string", "example": "drjoao@clinica.com" },
                "cro": { "type": "string", "example": "56789-SP" },
                "senha": { "type": "string", "example": "12345" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação bem sucedida"
          },
          "400": {
            "description": "Operação inválida"
          }
        }
      }
    },
...
    "/pacientes": {
      "get": {
        "tags": ["PACIENTES"],
        "summary": "Retorna pacientes",
        "description": "Retorna todos pacientes existentes",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Operação bem sucedida"
          }
        }
      },
      "post": {
        "tags": ["PACIENTES"],
        "summary": "Adiciona paciente",
        "description": "Adiciona uma paciente",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "paciente",
            "description": "Objeto que descreve um paciente",
            "schema": {
              "type": "object",
              "required": ["nome", "email", "senha", "fone"],
              "properties": {
                "nome": { "type": "string", "example": "Ana" },
                "senha": { "type": "string", "example": "12345" },
                "email": { "type": "string", "example": "ana@ana.com" },
                "fone": { "type": "string", "example": "(11) 9876-3455" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Inserção bem sucedida"
          }
        }
      }
    },
    "/pacientes/{id}": {
      "put": {
        "tags": ["PACIENTES"],
        "summary": "Altera paciente",
        "description": "Altera um paciente",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ID do paciente a ser alterada",
            "required": true
          },
          {
            "in": "body",
            "name": "paciente",
            "description": "Objeto que descreve um paciente",
            "schema": {
              "type": "object",
              "required": ["nome", "email", "senha", "fone"],
              "properties": {
                "nome": { "type": "string", "example": "Ana" },
                "senha": { "type": "string", "example": "12345" },
                "email": { "type": "string", "example": "ana@ana.com" },
                "fone": { "type": "string", "example": "(11) 9876-3455" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Inserção bem sucedida"
          }
        }
      },
      "delete": {
        "tags": ["PACIENTES"],
        "summary": "Remove paciente",
        "description": "Remove um paciente na lista de pacientes",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "ID do paciente a ser removido",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Remoção bem sucedida"
          },
          "400": {
            "description": "PACIENTE NÃO ENCONTRADO"
          }
        }
      }
    },
    "/profissionais": {
      "get": {
        "tags": ["PROFISSIONAIS"],
        "summary": "Retorna profissionais",
        "description": "Retorna todos profissionais existentes",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Operação bem sucedida"
          }
        }
      },
      "post": {
        "tags": ["PROFISSIONAIS"],
        "summary": "Adiciona profissional",
        "description": "Adiciona um profissional",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "profissional",
            "description": "Objeto que descreve um profissional",
            "schema": {
              "type": "object",
              "required": [
                "nome",
                "email",
                "senha",
                "fone",
                "especialidade",
                "descricao",
                "cro"
              ],
              "properties": {
                "nome": { "type": "string", "example": "Dr. João" },
                "senha": { "type": "string", "example": "12345" },
                "email": { "type": "string", "example": "drjoao@clinica.com" },
                "fone": { "type": "number", "example": "(11) 9876-3455" },
                "especialidade": {
                  "type": "string",
                  "example": "Endodontia"
                },
                "descricao": {
                  "type": "string",
                  "example": "Especialista em canais"
                },
                "cro": { "type": "string", "example": "56789-SP" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Inserção bem sucedida"
          }
        }
      }
    },
...
  }
}